+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-- Added support for creating documents containing UTF - fixes [Issue #151](https://github.com/PlagueHO/CosmosDB/issues/151).
+- Added support for creating and updating documents containing
+  non-ASCII characters by adding Encoding parameter to `New-CosmosDbDocument`
+  and `Set-CosmosDbDocument` functions - fixes [Issue #151](https://github.com/PlagueHO/CosmosDB/issues/151).
 - Fix table of contents link in README.MD.
 
 ## 2.1.9.92